+++ pkg/constant/errors.go
@@ -5,11 +5,36 @@ import "fmt"
 type ErrorCode int32
 
 const (
-	ErrCodeSuccess       ErrorCode = 0
-	ErrCodeInvalidParams ErrorCode = -1
+	ErrCodeSuccess        ErrorCode = 0
+	ErrCodeInvalidParams   ErrorCode = -1
 	ErrCodeTimeout       ErrorCode = -100
-	ErrCodeDisconnected   ErrorCode = -200
-	ErrCodeUnknown      ErrorCode = -400
+	ErrCodeDisconnected  ErrorCode = -200
+	ErrCodeUnknown       ErrorCode = -400
+
+	// Connection errors
+	ErrCodeNetworkError    ErrorCode = -101
+	ErrCodeProtocolErr  ErrorCode = -102
+	ErrCodeServerBusy    ErrorCode = -103
+
+	// Account errors
+	ErrCodeAccNotFound      ErrorCode = -201
+	ErrCodeAccDisabled    ErrorCode = -202
+	ErrCodeAccLocked      ErrorCode = -203
+	ErrCodeAccAuthFail    ErrorCode = -204
+
+	// Trading errors
+	ErrCodeInsufficientBalance ErrorCode = -301
+	ErrCodeMarketClosed      ErrorCode = -302
+	ErrCodeOrderRejected    ErrorCode = -303
+	ErrCodePriceOutOfRange ErrorCode = -304
+	ErrCodeQtyTooLarge      ErrorCode = -305
+	ErrCodeTradingDisabled ErrorCode = -306
+	ErrCodeInvalidSecurity ErrorCode = -307
+	ErrCodeNoPermission    ErrorCode = -308
+
+	// Subscription errors
+	ErrCodeAlreadySubbed  ErrorCode = -401
+	ErrCodeNotSubbed      ErrorCode = -402
 )
 
 type FutuError struct {
@@ -54,6 +79,63 @@ func IsInvalidParams(err error) bool {
 func IsSuccess(err error) bool {
 	if fe, ok := err.(*FutuError); ok {
 		return fe.Code == ErrCodeSuccess
+	}
+	return false
+}
+
+func getFutuError(err error) (*FutuError, bool) {
+	fe, ok := err.(*FutuError)
+	return fe, ok
+}
+
+func IsNetworkError(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeNetworkError || fe.Code == ErrCodeProtocolErr
+	}
+	return false
+}
+
+func IsServerBusy(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeServerBusy
+	}
+	return false
+}
+
+func IsAccountError(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		switch fe.Code {
+		case ErrCodeAccNotFound, ErrCodeAccDisabled, ErrCodeAccLocked, ErrCodeAccAuthFail:
+			return true
+		}
+	}
+	return false
+}
+
+func IsInsufficientBalance(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeInsufficientBalance
+	}
+	return false
+}
+
+func IsMarketClosed(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeMarketClosed
+	}
+	return false
+}
+
+func IsOrderRejected(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeOrderRejected
+	}
+	return false
+}
+
+func IsSubscriptionError(err error) bool {
+	if fe, ok := getFutuError(err); ok {
+		return fe.Code == ErrCodeAlreadySubbed || fe.Code == ErrCodeNotSubbed
 	}
 	return false
 }